+++ src/components/common/Form/error-message/index.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import Input from "../Input";
+import Input from "../input";
 import ErrorMessage from ".";
 
 const meta = {
