+++ src/components/common/drop-down/index.tsx
@@ -2,9 +2,9 @@ import { ReactNode } from "react";
 
 import useClickOutside from "@/hooks/useClickOutside";
 
-import DropDownItem from "./Item";
-import DropDownMenu from "./Menu";
-import DropDownTrigger from "./Trigger";
+import DropDownItem from "./item";
+import DropDownMenu from "./menu";
+import DropDownTrigger from "./trigger";
 
 interface DropDownProps {
   /** 드롭다운 컴포넌트 안에 포함될 내용입니다. */
